refactor(home, login-register  & product page) : restructure code by splitting components and grouping by pages
