+++ frontend/src/app/layout.tsx
@@ -1,28 +1,28 @@
-import type React from "react"
-import type { Metadata } from "next"
-import { Inter } from "next/font/google"
-import "./globals.css"
-import Header from "@/components/header"
-import Footer from "@/components/footer"
-import { Providers } from "@/components/providers"
+import { Providers } from "@/components/common/providers";
+import Footer from "@/layouts/home/footer";
+import Header from "@/layouts/home/header";
+import "@/styles/globals.css";
+import type { Metadata } from "next";
+import { Inter } from "next/font/google";
+import type React from "react";
 
 const inter = Inter({
   subsets: ["latin"],
   weight: ["400", "500", "600", "700"],
   variable: "--font-inter",
-})
+});
 
 export const metadata: Metadata = {
   title: "Nông Sàn - Sàn Thương Mại Điện Tử Nông Nghiệp",
   description:
     "Sàn thương mại điện tử chuyên về nông nghiệp, tập trung vào việc mua bán các sản phẩm nông nghiệp như rau củ, trái cây, hạt giống, vật tư nông nghiệp, và các sản phẩm chế biến từ nông nghiệp.",
-    generator: 'v0.dev'
-}
+  generator: "Huy Trần",
+};
 
 export default function RootLayout({
   children,
 }: {
-  children: React.ReactNode
+  children: React.ReactNode;
 }) {
   return (
     <html lang="vi" suppressHydrationWarning>
@@ -36,9 +36,5 @@ export default function RootLayout({
         </Providers>
       </body>
     </html>
-  )
+  );
 }
-
-
-
-import './globals.css'
