+++ frontend/src/app/seller/[id]/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import ProductCard from "@/components/product-card";
+import ProductCard from "@/components/products/product-card";
 import { Button } from "@/components/ui/button";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { products } from "@/data/products";
