+++ frontend/src/components/auth/LoginForm.tsx
@@ -1,0 +1,115 @@
+import { Alert, AlertDescription } from "@/components/ui/alert";
+import { Button } from "@/components/ui/button";
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardHeader,
+  CardTitle,
+} from "@/components/ui/card";
+import { Checkbox } from "@/components/ui/checkbox";
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
+import Link from "next/link";
+import type React from "react";
+import PasswordInput from "./PasswordInput";
+import SocialLoginButtons from "./SocialLoginButtons";
+
+interface LoginFormProps {
+  username: string;
+  setUsername: (value: string) => void;
+  password: string;
+  setPassword: (value: string) => void;
+  showPassword: boolean;
+  setShowPassword: (value: boolean) => void;
+  error: string | string[] | null;
+  loading: boolean;
+  handleSubmit: (e: React.FormEvent) => void;
+}
+
+export default function LoginForm({
+  username,
+  setUsername,
+  password,
+  setPassword,
+  showPassword,
+  setShowPassword,
+  error,
+  loading,
+  handleSubmit,
+}: LoginFormProps) {
+  const formatErrorMessage = (error: string | string[]) => {
+    if (Array.isArray(error)) {
+      return (
+        <ul className="list-disc pl-5">
+          {error.map((err, index) => (
+            <li key={index}>{err}</li>
+          ))}
+        </ul>
+      );
+    }
+    return error;
+  };
+
+  return (
+    <Card>
+      <CardHeader>
+        <CardTitle className="text-2xl text-center">Đăng nhập</CardTitle>
+        <CardDescription className="text-center">
+          Đăng nhập để tiếp tục mua sắm trên Nông Sàn
+        </CardDescription>
+      </CardHeader>
+      <CardContent>
+        {error && (
+          <Alert variant="destructive" className="mb-4">
+            <AlertDescription>{formatErrorMessage(error)}</AlertDescription>
+          </Alert>
+        )}
+        <form onSubmit={handleSubmit} className="space-y-4">
+          <div className="space-y-2">
+            <Label htmlFor="username">Username</Label>
+            <Input
+              id="username"
+              type="text"
+              placeholder="Nhập username"
+              value={username}
+              onChange={(e) => setUsername(e.target.value)}
+            />
+          </div>
+          <div className="space-y-2">
+            <div className="flex items-center justify-between">
+              <Label htmlFor="password">Mật khẩu</Label>
+              <Link
+                href="/forgot-password"
+                className="text-sm text-primary hover:underline"
+              >
+                Quên mật khẩu?
+              </Link>
+            </div>
+            <PasswordInput
+              id="password"
+              value={password}
+              onChange={(e) => setPassword(e.target.value)}
+              showPassword={showPassword}
+              toggleShowPassword={() => setShowPassword(!showPassword)}
+            />
+          </div>
+          <div className="flex items-center space-x-2">
+            <Checkbox id="remember" />
+            <Label htmlFor="remember" className="text-sm font-normal">
+              Ghi nhớ đăng nhập
+            </Label>
+          </div>
+          <Button
+            type="submit"
+            className="w-full bg-primary hover:bg-primary-dark"
+            disabled={loading}
+          >
+            {loading ? "Đang đăng nhập..." : "Đăng nhập"}
+          </Button>
+        </form>
+        <SocialLoginButtons />
+      </CardContent>
+    </Card>
+  );
+}
