+++ frontend/src/components/common/redux-provider.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+import { store } from "@/lib/store";
+import type React from "react";
+import { Provider } from "react-redux";
+
+export function ReduxProvider({ children }: { children: React.ReactNode }) {
+  return <Provider store={store}>{children}</Provider>;
+}
