+++ frontend/src/components/products/category-card.tsx
